feat: clear helapordo.h, helapordo.c
